fix: flutter analyze

+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/link_preview/shared.dart
@@ -2,23 +2,24 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_editor_plugins/appflowy_editor_plugins.dart';
 
 Future<void> convertUrlPreviewNodeToLink(
-    EditorState editorState, Node node) async {
-  assert(node.type == LinkPreviewBlockKeys.type);
+  EditorState editorState,
+  Node node,
+) async {
+  if (node.type != LinkPreviewBlockKeys.type) {
+    return;
+  }
+
   final url = node.attributes[ImageBlockKeys.url];
+  final delta = Delta()
+    ..insert(
+      url,
+      attributes: {
+        AppFlowyRichTextKeys.href: url,
+      },
+    );
   final transaction = editorState.transaction;
   transaction
-    ..insertNode(
-      node.path,
-      paragraphNode(
-        delta: Delta()
-          ..insert(
-            url,
-            attributes: {
-              AppFlowyRichTextKeys.href: url,
-            },
-          ),
-      ),
-    )
+    ..insertNode(node.path, paragraphNode(delta: delta))
     ..deleteNode(node);
   transaction.afterSelection = Selection.collapsed(
     Position(
